pyproject.toml
[tool.poetry]
name = "nashtaconstruction"
version = "0.1.0"
description = "This is a website for a construction company"
authors = ["Amir Mirfallahi"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
django-unfold = "^0.20.3"
pillow = "^10.2.0"
django-tinymce = "^3.7.1"
python-telegram-bot = "^20.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
